[781] Add nightly job to upsert hashed transactions
