+++ app/jobs/hashed_transaction_job/imports.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module HashedTransactionJob
+  class Imports < ApplicationJob
+    def perform
+      ::HashedTransactionService::PlaidTransaction::Import.new.run
+    end
+  end
+end
